+++ substrate-node/pallets/pallet-dao/src/mock.rs
@@ -8,8 +8,7 @@ use pallet_tfgrid::{
     farm::FarmName,
     interface::{InterfaceIp, InterfaceMac, InterfaceName},
     node::{Location, SerialNumber},
-    pub_config::{Domain, FullPublicIp6, GW4, GW6, IP4, IP6},
-    pub_ip::FullPublicIp4,
+    ip::{Domain, FullPublicIp4, FullPublicIp6, GW4, GW6, IP4, IP6},
     terms_cond::TermsAndConditions,
     twin::TwinIp,
     DocumentHashInput, DocumentLinkInput, PublicIpOf, TwinIpInput,
