+++ substrate-node/pallets/pallet-tfgrid/src/ip.rs
@@ -1,4 +1,4 @@
-use crate::{Config, DomainInput, Error, FullIp6Input, Gw4Input, Gw6Input, Ip4Input, Ip6Input};
+use crate::{Config, DomainInput, Error, FullIp6Input, FullIp4Input, Gw4Input, Gw6Input, Ip4Input, Ip6Input};
 use codec::{Decode, Encode, MaxEncodedLen};
 use frame_support::{
     ensure, sp_runtime::SaturatedConversion, traits::ConstU32, BoundedVec, RuntimeDebug,
@@ -268,6 +268,82 @@ fn validate_domain_name(input: &[u8]) -> bool {
     input
         .iter()
         .all(|c| matches!(c, b'a'..=b'z' | b'A'..=b'Z' | b'0'..=b'9' | b'-' | b'.'))
+}
+
+
+/// A Full IP4.
+/// Needs to be valid IP and Gateway, IP must also fall in range of gateway
+#[derive(Encode, Decode, Eq, RuntimeDebug, TypeInfo, MaxEncodedLen)]
+#[scale_info(skip_type_params(T))]
+#[codec(mel_bound())]
+pub struct FullPublicIp4<T: Config> {
+    pub ip: BoundedVec<u8, ConstU32<MAX_IP4_LENGTH>>,
+    pub gateway: BoundedVec<u8, ConstU32<MAX_GW4_LENGTH>>,
+    _marker: PhantomData<(T, ConstU32<MAX_GW4_LENGTH>)>,
+}
+
+impl<T: Config> TryFrom<FullIp4Input> for FullPublicIp4<T> {
+    type Error = Error<T>;
+
+    /// Fallible initialization from a provided byte vector if it is below the
+    /// minimum or exceeds the maximum allowed length or contains invalid ASCII
+    /// characters.
+    fn try_from(value: FullIp4Input) -> Result<Self, Self::Error> {
+        ensure!(
+            value.gw.len() >= MIN_GW4_LENGTH.saturated_into(),
+            Self::Error::GatewayIPTooShort
+        );
+        ensure!(
+            value.gw.len() <= MAX_GW4_LENGTH.saturated_into(),
+            Self::Error::GatewayIPTooLong
+        );
+
+        ensure!(
+            value.ip.len() >= MIN_IP4_LENGTH.saturated_into(),
+            Self::Error::PublicIPTooShort
+        );
+        ensure!(
+            value.ip.len() <= MAX_IP4_LENGTH.saturated_into(),
+            Self::Error::PublicIPTooLong
+        );
+
+        if let Ok(gw) = IPv4::parse(&value.gw) {
+            ensure!(gw.is_public() && gw.is_unicast(), Self::Error::InvalidGW4);
+            if let Ok(ip) = IPv4Cidr::parse(&value.ip) {
+                ensure!(
+                    ip.is_public() && ip.is_unicast() && ip.contains(gw),
+                    Self::Error::InvalidIP4
+                );
+            } else {
+                return Err(Self::Error::InvalidIP4);
+            }
+            Ok(Self {
+                ip: value.ip,
+                gateway: value.gw,
+                _marker: PhantomData,
+            })
+        } else {
+            Err(Self::Error::InvalidGW4)
+        }
+    }
+}
+
+// FIXME: did not find a way to automatically implement this.
+impl<T: Config> PartialEq for FullPublicIp4<T> {
+    fn eq(&self, other: &Self) -> bool {
+        self.ip == other.ip && self.gateway == other.gateway
+    }
+}
+
+// FIXME: did not find a way to automatically implement this.
+impl<T: Config> Clone for FullPublicIp4<T> {
+    fn clone(&self) -> Self {
+        Self {
+            ip: self.ip.clone(),
+            gateway: self.gateway.clone(),
+            _marker: PhantomData,
+        }
+    }
 }
 
 /// A Full IP4.
